Rework native surfaces for flexibility and simplicity
Currently, the Linux native surface can be either a memory-buffer
surface or a Pixmap, chosen at runtime. Since the memory-buffer surface
is platform-independent, we push this configurability to the top-level
native surface API. This will allow having multiple surface types
per-platform, which is useful for both adding Wayland support and
adding configurable surface transports for performance tuning. This
also simplifies the surface interface a great deal.

The other reason for this change is it prepares the rust-layers
implementation for handling more of the glue between Azure, skia, and
Servo, so that the first two do not need intimate details of the Servo
compositor. This will also make it easier to rework how DrawTargets in
Azure interact with rust-layers native surfaces.