Reuse MemoryBufferNativeSurface for Android
Android native surfaces are either memory buffers or EGLImages.
Currently the Android memory buffer surface implementation exists in
addition to the MemoryBufferNativeSurface which is unused for Android.
This change makes the Android native surface an enum that can either
hold a MemoryBufferNativeSurface or an EGLImage-based native surface.
The end result is that there is now only one implementation of memory
buffer surfaces shared among all platforms.